move filter component to public folder
